重构数据模型框架,提取分区对象基类.
