+++ src/test/java/cn/easyxue/mybatis/shards/domain/AbstractShardEntity.java
@@ -9,49 +9,11 @@
 package cn.easyxue.mybatis.shards.domain;
 
 import java.io.Serializable;
+import org.makersoft.shards.ShardedEntity;
 
-import org.makersoft.shards.annotation.PrimaryKey;
 import org.makersoft.shards.annotation.ShardEntity;
 
 @ShardEntity
-public abstract class AbstractShardEntity implements Serializable{
-
-	private static final long serialVersionUID = 5765444163885433060L;
-
-	@PrimaryKey
-	protected String id;
-
-	public final String getId() {
-		return id;
-	}
-
-	public void setId(String id) {
-		this.id = id;
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((id == null) ? 0 : id.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		AbstractShardEntity other = (AbstractShardEntity) obj;
-		if (id == null) {
-			if (other.id != null)
-				return false;
-		} else if (!id.equals(other.id))
-			return false;
-		return true;
-	}
+public abstract class AbstractShardEntity extends AbstractEntity implements Serializable,ShardedEntity {
 
 }
