+++ src/test/java/cn/easyxue/mybatis/shards/domain/shard0/Company.java
@@ -8,48 +8,49 @@
  */
 package cn.easyxue.mybatis.shards.domain.shard0;
 
-import cn.easyxue.mybatis.shards.domain.AbstractShardEntity;
+import cn.easyxue.mybatis.shards.domain.AbstractEntity;
 
 /**
  * entity role for test.
  */
-public class Company extends AbstractShardEntity {
-	private static final long serialVersionUID = 2960427786748550511L;
+public class Company extends AbstractEntity {
 
-	private String name;
-	private String creatorId;
+    private static final long serialVersionUID = 2960427786748550511L;
+
+    private String name;
+    private String creatorId;
     /**
      * 数据分区名.
      */
     private String dbKey;
-    
-    
-	public Company() {
-	}
 
-	public Company(String name, String code) {
-		this.name = name;
-		this.creatorId = code;
-	}
+    public Company() {
+    }
 
-	public String getName() {
-		return name;
-	}
+    public Company(String name, String code) {
+        this.name = name;
+        this.creatorId = code;
+    }
 
-	public void setName(String name) {
-		this.name = name;
-	}
+    public String getName() {
+        return name;
+    }
 
-	public String getCreatorId() {
-		return creatorId;
-	}
+    public void setName(String name) {
+        this.name = name;
+    }
 
-	public void setCreatorId(String creatorId) {
-		this.creatorId = creatorId;
-	}
+    public String getCreatorId() {
+        return creatorId;
+    }
+
+    public void setCreatorId(String creatorId) {
+        this.creatorId = creatorId;
+    }
 
     /**
      * 数据分区名.
+     *
      * @return the dbKey
      */
     public String getDbKey() {
@@ -58,6 +59,7 @@ public class Company extends AbstractShardEntity {
 
     /**
      * 数据分区名.
+     *
      * @param dbKey the dbKey to set
      */
     public void setDbKey(String dbKey) {
@@ -69,7 +71,7 @@ public class Company extends AbstractShardEntity {
      */
     public void build() {
         //员工表.
-        
+
     }
 
 }
